reddit/requirements.txt
boto3
pandas
praw
python-dotenv